// api/src/domains/dealer.domain.ts
export interface IDealer {
    id: number;
    name: string;
    address: string;
    phone: string;
    email: string;
    tax_id: string;
    tax_number: string;
    tax_address: string;
    tax_phone: string;
    tax_email: string;
    created_at: Date;
    updated_at: Date;
    deleted_at: Date;
}